python_service/requirements.txt
flask
flask-cors
tensorflow
keras
opencv-python
mediapipe
numpy
pillow
scikit-learn